+++ core/src/main/scala/cats/Applicative.scala
@@ -1,6 +1,6 @@
 package cats
 
-import cats.std.list._
+import cats.instances.list._
 import simulacrum.typeclass
 
 /**
@@ -33,8 +33,8 @@ import simulacrum.typeclass
     ap(pure(f))(fa)
 
   /**
-    * Given `fa` and `n`, apply `fa` `n` times to construct an `F[List[A]]` value.
-    */
+   * Given `fa` and `n`, apply `fa` `n` times to construct an `F[List[A]]` value.
+   */
   def replicateA[A](n: Int, fa: F[A]): F[List[A]] =
     sequence(List.fill(n)(fa))