+++ core/src/main/scala/cats/syntax/bifunctor.scala
@@ -10,7 +10,7 @@ trait BifunctorSyntax {
 }
 
 final class BifunctorOps[F[_, _], A, B](fab: F[A, B])(implicit F: Bifunctor[F]) {
-  def bimap[C, D](f: A => C, g: B => D): F[C,D] = F.bimap(fab)(f,g)
+  def bimap[C, D](f: A => C, g: B => D): F[C, D] = F.bimap(fab)(f, g)
 
   def leftMap[C](f: A => C): F[C, B] = F.leftMap(fab)(f)